+++ ubelt/util_platform.py
@@ -137,6 +137,11 @@ def get_app_data_dir(appname, *args):
     SeeAlso:
         :func:`ensure_app_data_dir`
     """
+    import ubelt as ub
+    ub.schedule_deprecation(
+        modname='ubelt', name='get_app_data_dir and ensure_app_data_dir', type='function',
+        migration='use ubelt.Path.appdir(type="data") instead',
+        deprecate='1.2.0', error='2.0.0', remove='2.1.0')
     dpath = join(platform_data_dir(), appname, *args)
     return dpath
 
@@ -189,6 +194,11 @@ def get_app_config_dir(appname, *args):
     SeeAlso:
         :func:`ensure_app_config_dir`
     """
+    import ubelt as ub
+    ub.schedule_deprecation(
+        modname='ubelt', name='get_app_config_dir and ensure_app_config_dir', type='function',
+        migration='use ubelt.Path.appdir(type="config") instead',
+        deprecate='1.2.0', error='2.0.0', remove='2.1.0')
     dpath = join(platform_config_dir(), appname, *args)
     return dpath
 
@@ -244,6 +254,11 @@ def get_app_cache_dir(appname, *args):
     SeeAlso:
         :func:`ensure_app_cache_dir`
     """
+    import ubelt as ub
+    ub.schedule_deprecation(
+        modname='ubelt', name='get_app_cache_dir and ensure_app_cache_dir', type='function',
+        migration='use ubelt.Path.appdir(type="cache") instead',
+        deprecate='1.2.0', error='2.0.0', remove='2.1.0')
     dpath = join(platform_cache_dir(), appname, *args)
     return dpath